requirements.txt
fastapi==0.97.0
pydantic==1.10.7
python-dotenv==1.0.0
Requests==2.31.0
websocket_client==1.5.1
